Increment 5 completed
Added env config
